+++ core/site.ts
@@ -18,9 +18,10 @@ import Events from "./events.ts";
 import Formats from "./formats.ts";
 import Searcher from "./searcher.ts";
 import Scripts from "./scripts.ts";
-import FSWatcher from "../core/watcher.ts";
+import Archetypes from "./archetypes.ts";
+import FSWatcher from "./watcher.ts";
 import { FSWriter } from "./writer.ts";
-import { filesToPages, Page } from "./file.ts";
+import { filesToPages, Page, StaticFile } from "./file.ts";
 import textLoader from "./loaders/text.ts";
 import binaryLoader from "./loaders/binary.ts";
 import Server from "./server.ts";
@@ -28,10 +29,11 @@ import Cache from "./cache.ts";
 import DebugBar from "./debugbar.ts";
 import notFound from "../middlewares/not_found.ts";
 
+import type { Archetype } from "./archetypes.ts";
 import type { Entry, Loader } from "./fs.ts";
 import type { BasenameParser, Destination } from "./source.ts";
 import type { Components, UserComponent } from "./components.ts";
-import { Data, RawData, StaticFile } from "./file.ts";
+import type { Data, RawData } from "./file.ts";
 import type { Engine, Helper, HelperOptions, HelperThis } from "./renderer.ts";
 import type { Event, EventListener, EventOptions } from "./events.ts";
 import type { Processor } from "./processors.ts";
@@ -80,6 +82,9 @@ export default class Site {
 
   /** Internal data. Used to save arbitrary data by plugins and processors */
   _data: Record<string, unknown> = {};
+
+  /** To register and run archetypes */
+  archetypes: Archetypes;
 
   /** To read the files from the filesystem */
   fs: FS;
@@ -154,6 +159,7 @@ export default class Site {
   constructor(options?: SiteOptions) {
     this.options = merge(defaults, options);
 
+    const root = this.root();
     const src = this.src();
     const dest = this.dest();
     const { includes, cwd, prettyUrls, components, server, caseSensitiveUrls } =
@@ -194,6 +200,7 @@ export default class Site {
     });
 
     // Other stuff
+    const archetypes = new Archetypes({ src, root });
     const events = new Events<SiteEvent>();
     const scripts = new Scripts({ cwd });
     const writer = new FSWriter({ dest, caseSensitiveUrls });
@@ -209,6 +216,7 @@ export default class Site {
     });
 
     // Save everything in the site instance
+    this.archetypes = archetypes;
     this.fs = fs;
     this.formats = formats;
     this.componentLoader = componentLoader;
@@ -465,6 +473,12 @@ export default class Site {
     const pages = this.scopedPages.get(scope) || [];
     pages.push(data);
     this.scopedPages.set(scope, pages);
+    return this;
+  }
+
+  /** Register an archetype */
+  archetype(name: string, archetype: string | Archetype): this {
+    this.archetypes.register(name, archetype);
     return this;
   }