+++ types.ts
@@ -2,7 +2,7 @@ import type { Engine, Helper } from "./core/renderer.ts";
 import type { Data as PageData, Page } from "./core/file.ts";
 import type { Loader } from "./core/fs.ts";
 import type { default as Site, Plugin } from "./core/site.ts";
-import type { Archetype, ArchetypeFile } from "./cli/create.ts";
+import type { Archetype, ArchetypeFile } from "./core/archetypes.ts";
 import type { Middleware, RequestHandler } from "./core/server.ts";
 
 declare global {